# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalItem": {
        "additionalProperties": false,
        "description": "One artefact awaiting a decision.",
        "properties": {
          "artefactSha256": {
            "description": "The bytes being decided on.",
            "title": "Artefactsha256",
            "type": "string"
          },
          "awaitingSince": {
            "description": "When it entered the queue.",
            "format": "date-time",
            "title": "Awaitingsince",
            "type": "string"
          },
          "gates": {
            "description": "The gate results.",
            "items": {
              "$ref": "#/components/schemas/GateOut"
            },
            "title": "Gates",
            "type": "array"
          },
          "id": {
            "description": "The subject awaiting approval.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "description": "What the artefact is.",
            "title": "Model",
            "type": "string"
          },
          "runId": {
            "description": "The run that produced it.",
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          },
          "submittedBy": {
            "description": "Who submitted the run.",
            "title": "Submittedby",
            "type": "string"
          }
        },
        "required": [
          "id",
          "runId",
          "model",
          "artefactSha256",
          "submittedBy",
          "awaitingSince"
        ],
        "title": "ApprovalItem",
        "type": "object"
      },
      "ApprovalPage": {
        "additionalProperties": false,
        "description": "A page of the approval queue.",
        "properties": {
          "items": {
            "description": "This page's pending approvals.",
            "items": {
              "$ref": "#/components/schemas/ApprovalItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Page size.",
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cursor for the next page.",
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "limit"
        ],
        "title": "ApprovalPage",
        "type": "object"
      },
...
      "DecisionIn": {
        "additionalProperties": false,
        "description": "An approver's signed decision. SAD 7.1 `approval`.",
        "properties": {
          "decision": {
            "description": "The decision taken.",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "description": "Why. Recorded on the approval and visible in lineage.",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "signature": {
            "description": "Detached signature over the approval payload, from the approver's key. Whitespace is stripped and an empty result is refused: a blank signature is a row somebody could have written, and publication verifies it.",
            "minLength": 1,
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason",
          "signature"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
      "DecisionOut": {
        "additionalProperties": false,
        "description": "A recorded decision.",
        "properties": {
          "approver": {
            "description": "Who decided.",
            "title": "Approver",
            "type": "string"
          },
          "decidedAt": {
            "description": "When it was decided.",
            "format": "date-time",
            "title": "Decidedat",
            "type": "string"
          },
          "decision": {
            "description": "The decision taken.",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "description": "The approval record.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "description": "Why.",
            "title": "Reason",
            "type": "string"
          },
          "soleApproverException": {
            "description": "True where the approver also submitted the run. Computed, never supplied. Constraint C-11; visible here, in the lineage and in the model card.",
            "title": "Soleapproverexception",
            "type": "boolean"
          },
          "subjectId": {
            "description": "What was decided on.",
            "format": "uuid",
            "title": "Subjectid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectId",
          "approver",
          "decision",
          "reason",
          "soleApproverException",
          "decidedAt"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
...
      "GateOut": {
        "additionalProperties": false,
        "description": "One gate result, with its baseline and margin. SAD 7.1.",
        "properties": {
          "baselineValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "What it was compared against.",
            "title": "Baselinevalue"
          },
          "gate": {
            "description": "Gate identifier, E1 to E6.",
            "title": "Gate",
            "type": "string"
          },
          "margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Measurement minus baseline.",
            "title": "Margin"
          },
          "passed": {
            "description": "Whether the gate was satisfied.",
            "title": "Passed",
            "type": "boolean"
          },
          "suiteVersion": {
            "description": "The suite version that produced it.",
            "title": "Suiteversion",
            "type": "string"
          },
          "value": {
            "description": "The measurement.",
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "gate",
          "suiteVersion",
          "value",
          "passed"
        ],
        "title": "GateOut",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem document.",
        "properties": {
          "code": {
            "description": "Machine readable, stable problem code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation for this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URI of this occurrence.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "PublishOut": {
        "additionalProperties": false,
        "description": "What a publication produced.",
        "properties": {
          "artefactSha256": {
            "description": "The bytes that were published.",
            "title": "Artefactsha256",
            "type": "string"
          },
          "formats": {
            "description": "Which formats were published.",
            "items": {
              "type": "string"
            },
            "title": "Formats",
            "type": "array"
          },
          "manifest": {
            "additionalProperties": true,
            "description": "The SHA-256 manifest of the release package.",
            "title": "Manifest",
            "type": "object"
          },
          "model": {
            "description": "The released model.",
            "title": "Model",
            "type": "string"
          },
          "releasedAt": {
            "description": "When it was published.",
            "format": "date-time",
            "title": "Releasedat",
            "type": "string"
          }
        },
        "required": [
          "artefactSha256",
          "model",
          "releasedAt",
          "formats",
          "manifest"
        ],
        "title": "PublishOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nThe DRAUPNIR control plane API.\n\nErrors are RFC 9457 problem documents served as `application/problem+json`.\nMutating endpoints require an `Idempotency-Key`. Collections are cursor\npaginated. Mutable resources carry an `ETag` and require `If-Match`.\nLong operations return 202 with a run identifier; nothing blocks on training.\n",
    "summary": "CIM-56 model factory control plane",
    "title": "DRAUPNIR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/gates": {
      "get": {
        "description": "List artefacts awaiting a decision, with their gate results.\n\n    The gate results come back with the queue rather than behind a second\n    request per row. AC-U13 puts the evidence above the decision control, and a\n    queue that has to be expanded row by row to see any of it is a queue whose\n    evidence is, in practice, after the decision.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "listGates",
        "parameters": [
          {
            "description": "Opaque cursor from a previous page's `nextCursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's `nextCursor`.",
              "title": "Cursor"
            }
          },
          {
            "description": "Which part of the queue to return.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "pending",
              "description": "Which part of the queue to return.",
              "enum": [
                "pending",
                "decided",
                "all"
              ],
              "title": "State",
              "type": "string"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size.",
              "title": "Limit"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "The approval queue",
        "tags": [
          "approvals"
        ]
      }
    },
    "/v1/gates/{gate_id}/decide": {
      "post": {
        "description": "Record a signed decision.\n\n    The sole approver exception is computed here from approver against\n    submitter and is never accepted from the request. Constraint C-11: there is\n    no argument that sets it, so suppressing it means editing GLEIPNIR, which\n    is a code review rather than a deployment.\n    \n\nRequires: `approver`.",
        "operationId": "decideGate",
        "parameters": [
          {
            "description": "The subject awaiting a decision.",
            "in": "path",
            "name": "gate_id",
            "required": true,
            "schema": {
              "description": "The subject awaiting a decision.",
              "format": "uuid",
              "title": "Gate Id",
              "type": "string"
            }
          },
          {
            "description": "Entity tag the write is conditional on.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity tag the write is conditional on.",
              "title": "If-Match"
            }
          },
          {
            "description": "Replaying a request with the same key returns the original result.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replaying a request with the same key returns the original result.",
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Approve or reject",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/v1/releases/{artefact}/publish": {
      "post": {
        "description": "Publish, having re-verified the artefact and its approval.\n\n    Every refusal in `skidbladnir.publish` applies: the bytes are re-hashed and\n    compared against the gate evidence (AC-S8), every built format must have\n    passing evidence (AC-F9), the approval must be present and signed, and the\n    federation must have countersigned an anchor at or beyond this release's\n    sequence (AC-S13).\n    \n\nRequires: `approver`.",
        "operationId": "publishRelease",
        "parameters": [
          {
            "description": "The artefact's SHA-256. Publication re-verifies it (AC-S8).",
            "in": "path",
            "name": "artefact",
            "required": true,
            "schema": {
              "description": "The artefact's SHA-256. Publication re-verifies it (AC-S8).",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Artefact",
              "type": "string"
            }
          },
          {
            "description": "Entity tag the write is conditional on.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity tag the write is conditional on.",
              "title": "If-Match"
            }
          },
          {
            "description": "Replaying a request with the same key returns the original result.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replaying a request with the same key returns the original result.",
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Publish a release",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}